add company role and policy

+++ app/Http/Controllers/CompanyUserController.php
@@ -8,11 +8,14 @@ use App\Enums\Role;
 use App\Models\User;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
+use Illuminate\Support\Facades\Gate;
 
 class CompanyUserController extends Controller
 {
     public function index(Company $company)
     {
+        Gate::authorize('viewAny', $company); 
+
         $users = $users = $company->users()->where('role_id', Role::COMPANY_OWNER->value)->get();
  
         return view('companies.users.index', compact('company', 'users'));
@@ -20,11 +23,15 @@ class CompanyUserController extends Controller
 
     public function create(Company $company)
     {
+        Gate::authorize('create', $company);
+
         return view('companies.users.create', compact('company'));
     }
  
     public function store(StoreUserRequest $request, Company $company)
     {
+        Gate::authorize('create', $company); 
+
         $company->users()->create([
             'name' => $request->input('name'),
             'email' => $request->input('email'),
@@ -37,11 +44,15 @@ class CompanyUserController extends Controller
  
     public function edit(Company $company, User $user)
     {
+        Gate::authorize('update', $company); 
+
         return view('companies.users.edit', compact('company', 'user'));
     }
  
     public function update(UpdateUserRequest $request, Company $company, User $user)
     {
+        Gate::authorize('update', $company); 
+
         $user->update($request->validated());
  
         return to_route('companies.users.index', $company);
@@ -49,6 +60,8 @@ class CompanyUserController extends Controller
 
     public function destroy(Company $company, User $user)
     {
+        Gate::authorize('delete', $company); 
+
         $user->delete();
  
         return to_route('companies.users.index', $company);
